\name{HDLSSkST-package}
\alias{HDLSSkST-package}
\docType{package}
\title{
    Distribution-Free Exact High Dimensional Low Sample Size k-Sample Tests
}
\description{
This package performs some k-sample tests in high dimensional low sample size (HDLSS)
             regime based on clustering. This package includes four tests, namely (i) RI test, (ii) FEI test, (iii) ARI test, and (iv) AFEI test. These tests are distribution free, and exact level (size) \eqn{\alpha} test, also consistent as dimension \eqn{d} goes to \eqn{\infty}. We also apply Holm's step-down-procedure (1979) for multiple testing.
}
\author{
Biplab Paul, Shyamal K. De and Anil K. Ghosh

Maintainer: Biplab Paul<biplab.paul@niser.ac.in>
}
\references{
Biplab Paul, Shyamal K De and Anil K Ghosh (2020).  Distribution-free Exact k-sample Tests for High Dimensional Low Sample Size data Based on Clustering.

Soham Sarkar and Anil K Ghosh (2019). On perfect clustering of high dimension, low sample size data, \emph{IEEE transactions on pattern analysis and machine intelligence}, doi:10.1109/TPAMI.2019.2912599.

William M Rand (1971). Objective criteria for the evaluation of clustering methods, \emph{Journal of the American Statistical association}, 66(336):846-850, doi:10.1080/01621459.1971.10482356.

Cyrus R Mehta and Nitin R Patel (1983). A network algorithm for performing Fisher's exact test in rxc contingency tables, \emph{Journal of the American Statistical Association}, 78(382):427-434, doi:10.2307/2288652.

Joseph C Dunn (1973). A fuzzy relative of the isodata process and its use in detecting compact well-separated clusters, doi:10.1080/01969727308546046.

Sture Holm (1979). A simple sequentially rejective multiple test procedure, \emph{Scandinavian journal of statistics}, 65-70, doi:10.2307/4615733.
}
\keyword{ package }
\keyword{ HDLSS}
